+++ dgdg-api/src/main/java/com/depromeet/dgdg/service/token/dto/JwtProperties.kt
@@ -1,11 +1,12 @@
-package com.depromeet.dgdg.service.token.property
+package com.depromeet.dgdg.service.token.dto
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.boot.context.properties.ConstructorBinding
 
 @ConstructorBinding
 @ConfigurationProperties("jwt")
-data class JwtProperty(
+data class JwtProperties(
     val issuer: String,
     val secret: String,
+    val expiresTime: Long
 )
